requirements.txt
# Gemma 4 MCP Dependencies
# MLX Framework
mlx-lm>=0.18.0

# MCP SDK
mcp>=1.0.0

# Tools
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Optional: より高度なスクレイピング (将来的にPlaywright統合時)
# playwright>=1.40.0
